requirements.txt
alembic==1.3.3
astroid==2.3.3
Babel==2.8.0
Click==7.0
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Babel==0.12.2
Flask-BabelEx==0.9.4
Flask-Migrate==2.5.2
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
gevent==20.5.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
psycopg2-binary==2.8.5
pylint==2.4.4
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2019.3
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.13
toml==0.10.0
typed-ast==1.4.1
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1

